chore: split `SystemStatus.tsx` into multiple files
