+++ packages/webui/src/client/ui/Settings/components/GenericDeviceSettingsComponent.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { PeripheralDevice, PeripheralDeviceType } from '@sofie-automation/corelib/dist/dataModel/PeripheralDevice'
-import { DeviceItem } from '../../Status/SystemStatus'
+import { DeviceItem } from '../../Status/SystemStatus/DeviceItem'
 import { ConfigManifestOAuthFlowComponent } from './ConfigManifestOAuthFlow'
 import { protectString, unprotectString } from '../../../lib/tempLib'
 import { SchemaFormForCollection } from '../../../lib/forms/SchemaFormForCollection'
